Support getting metric-specific CIDs
The metric name can be passed to the `cids` method to support logic
depending on the metric name. Since the method is used to allow getting
all CIDs that are required to calculate certain component type metrics,
the CIDs can differ depending on the metric that is calculated.

Without passing a metric, the logic of the method is unchanged. If a
metric was passed, the method extracts the CIDs from the formula. The
formula can be either a user-defined formula or a default formula if
available.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>