Add module for microgrid component data
Introduces the MicrogridData class which brings functionality to
determine power and SoC data aggregate over multiple components
of the same type. The mapping which components belong to the same
type has to be provided by the user via a config file.
It uses the reporting API client to fetch data.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>