Add data fetching for asset optimization
Add helper functions to initialize an microgrid data instance from
environment variables. Those can exist in the environment or optionally
be specified in an env file that can be passed to the function.

In addition to that a function for fetching and preparing the data as
required for the visualization functions is added.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>