Add matplotlib plotting for asset optimization
A function that can be used to visualize the power flow between
batteries, power generating and consuming assets in a stacked view is
provided. A second function can be used to plot the excess and deficit
net power, battery charge and battery SoC. Finally functionality to plot
monthly aggregates is added.

All functions in this module are based on matplotlib.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>