Revert "Widen weather API client dependency range (#145)"
This reverts commit 33fe0fe. `v0.3.0` of the weather client has a breaking
change which affects the module `frequenz.datasci.weather.weather_api.py`
In particular, the client method `hist_forecast_iterator` is replaced by
the `async` method `stream_historical_forecast`. Also, `v0.3.0` uses a
new version of the API, that the weather service doesn't support yet.
When the service supports it, the client dependency should be updated
to minimum `v0.3.0` due to the breaking change.

Signed-off-by: cyiallou - Costas <42914163+cyiallou@users.noreply.github.com>