Remove component type field from ComponentTypeConfig
This is redundant with the key of the dict entry. We may introduce it
back later but at the moment it's not set in the config files.

This also removes the default formula for battery SoC. This is intended as
it's preferred that users explicitly set a formula for this metric
instead of defaulting to a formula that only works for very specific
microgrids (namely where all battery components have the same effective
capacity).

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>